requirements.txt
python-dotenv>=1.0.1,<2
PyYAML>=6.0.1,<7
requests>=2.32.3,<3

